changed handicap_index column from integer to float in players model

+++ db/migrate/20141016235333_add_handicap_index_to_players.rb
@@ -1,5 +1,5 @@
 class AddHandicapIndexToPlayers < ActiveRecord::Migration
   def change
-    add_column :players, :handicap_index, :integer
+    add_column :players, :handicap_index, :float
   end
 end
